Split when and then steps into modules
